+++ qml/ControllerDialog.qml
@@ -58,8 +58,14 @@ WalletDialog {
                 return
             }
             if (resolver instanceof SignTransactionResolver) {
-                stack_view.push(sign_transaction_resolver_view_component, { resolver })
-                resolver.resolve()
+                if (resolver.device instanceof JadeDevice) {
+                    const view = jade_sign_transaction_view.createObject(stack_view, { resolver, wallet })
+                    stack_view.push(view)
+                    resolver.resolve()
+                } else {
+                    stack_view.push(sign_transaction_resolver_view_component, { resolver })
+                    resolver.resolve()
+                }
                 return
             }
             if (resolver instanceof SignLiquidTransactionResolver) {
@@ -97,6 +103,11 @@ WalletDialog {
     Component {
         id: sign_liquid_transaction_resolver_view_component
         SignLiquidTransactionResolverView {}
+    }
+
+    Component {
+        id: jade_sign_transaction_view
+        JadeSignTransactionView {}
     }
 
     Component {