+++ src/resolver.cpp
@@ -118,6 +118,8 @@ void GetXPubsResolver::resolve()
 
 SignTransactionResolver::SignTransactionResolver(Handler* handler, Device* device, const QJsonObject& result)
     : DeviceResolver(handler, device, result)
+    , m_transaction(m_required_data.value("transaction").toObject())
+    , m_outputs(m_required_data.value("transaction_outputs").toArray())
 {
 }