+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Enable Jade login button after logout
 
 ## [1.0.4] - 2021-11-19
 ### Added